This html file includes the steps we need to take and a start code for reading one month's data
This code includes explanation on which parameters to keep from the original csv file and which ones to add. In addition, it provides the steps we need to start doing some data visualization.
I added a basic starter code that reads the url csv file form the website and stores it as a dataframe for one month. I also added a code to convert the times to datetime format to calculate trip lengths in seconds.